Reorganize to eliminate some of the redundant code
